# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/ewhite1/product-api/handlers
...
info:
  description: Documentation for Product API
  title: of Product API
  version: 1.0.0
paths:
...
  /products/{id}:
    delete:
      description: Delete a products details
      operationId: deleteProduct
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
...
  noContentResponse:
    description: No content is returned by this API endpoint
...
schemes:
- http
swagger: "2.0"